requirements.txt
fastapi
uvicorn
cohere>=4.0.0
langchain>=0.1.16
langchain-community>=0.0.36
faiss-cpu
python-dotenv
